requirements.txt
amqp==5.1.0
asgiref==3.5.0
autoflake==1.4
billiard==3.6.4.0
boto3==1.21.27
botocore==1.24.27
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==4.0.3
django-storages==1.12.3
dropbox==11.28.0
gunicorn==20.1.0
idna==3.3
jmespath==1.0.0
kombu==5.2.4
lxml==4.8.0
ply==3.11
prompt-toolkit==3.0.28
pyflakes==2.4.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.1
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
stone==3.3.1
tzdata==2022.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
